# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login dengan email & password",
        "description": "Login pakai kredensial Supabase Auth. Mengembalikan access_token untuk dipakai di endpoint cms/. Salin access_token, klik Authorize di atas, tempel di sana.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "access_token": "eyJhbGci...",
                      "refresh_token": "v1.Mr...",
                      "expires_in": 3600,
                      "token_type": "bearer",
                      "user": {
                        "id": "uuid",
                        "email": "admin@nawasena.id",
                        "role": "admin"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Email atau password salah"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Perbarui access_token",
        "description": "Pakai refresh_token untuk dapat access_token baru (token lama berlaku 1 jam).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Profil user dari token",
        "description": "Mengembalikan id, email, dan role dari access_token yang dikirim.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "id": "uuid",
                      "email": "admin@nawasena.id",
                      "role": "admin"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/v1/cms/harvest": {
      "get": {
        "operationId": "HarvestCmsController_listAll",
        "summary": "Semua data panen (admin)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/harvest"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "HarvestCmsController_create",
        "summary": "Input data panen",
        "description": "Yield ton/ha & estimasi total dihitung otomatis dari ubinan (faktor BPS 1.6). Jika scheduleId diisi, jadwal ditandai harvested.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHarvestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/harvest"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/harvest/{id}": {
      "get": {
        "operationId": "HarvestCmsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/harvest"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "HarvestCmsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHarvestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/harvest"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "HarvestCmsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/harvest"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NAWASENA Center API",
    "description": "Backend NAWASENA Center untuk PPK Ormawa RDC UAD.\n\n## Cara pakai\n- Endpoint **public/** bisa diakses tanpa login (untuk website publik).\n- Endpoint **cms/** butuh login admin. Klik tombol **Authorize** di kanan atas, masukkan token (format: `Bearer <token>` tidak perlu, cukup token-nya).\n\n## Format response\nSemua response: `{ success: true, data: {...} }` atau `{ success: false, error: {...} }`.\n\n## Autentikasi\nLogin via Supabase Auth untuk dapat access_token, lalu Authorize di sini.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Masukkan access_token dari Supabase Auth (tanpa kata Bearer)"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@nawasena.id",
            "description": "Email akun (terdaftar di Supabase Auth)"
          },
          "password": {
            "type": "string",
            "example": "admin1234",
            "description": "Password akun"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "v1.Mr...",
            "description": "refresh_token dari hasil login"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
...
      "CreateHarvestDto": {
        "type": "object",
        "properties": {
          "farmerName": {
            "type": "string",
            "example": "Pak Sutrisno",
            "description": "Nama petani"
          },
          "fieldName": {
            "type": "string",
            "example": "Sawah Blok A"
          },
          "padukuhan": {
            "type": "string",
            "example": "Karangtengah",
            "description": "Padukuhan"
          },
          "ubinanKg": {
            "type": "number",
            "example": 4.2,
            "description": "Berat hasil ubinan 2.5x2.5m (kg)"
          },
          "areaHa": {
            "type": "number",
            "example": 0.5,
            "description": "Luas lahan petani (hektar) untuk estimasi total"
          },
          "pestLossPct": {
            "type": "number",
            "example": 8.5,
            "description": "Persen kerusakan akibat hama (target < 15%)"
          },
          "harvestDate": {
            "type": "string",
            "example": "2026-10-15",
            "description": "Tanggal panen"
          },
          "scheduleId": {
            "type": "string",
            "example": "uuid-jadwal",
            "description": "ID jadwal tanam terkait (opsional)"
          }
        },
        "required": [
          "farmerName",
          "ubinanKg",
          "areaHa",
          "harvestDate"
        ]
      },
      "UpdateHarvestDto": {
        "type": "object",
        "properties": {
          "farmerName": {
            "type": "string",
            "example": "Pak Sutrisno",
            "description": "Nama petani"
          },
          "fieldName": {
            "type": "string",
            "example": "Sawah Blok A"
          },
          "padukuhan": {
            "type": "string",
            "example": "Karangtengah",
            "description": "Padukuhan"
          },
          "ubinanKg": {
            "type": "number",
            "example": 4.2,
            "description": "Berat hasil ubinan 2.5x2.5m (kg)"
          },
          "areaHa": {
            "type": "number",
            "example": 0.5,
            "description": "Luas lahan petani (hektar) untuk estimasi total"
          },
          "pestLossPct": {
            "type": "number",
            "example": 8.5,
            "description": "Persen kerusakan akibat hama (target < 15%)"
          },
          "harvestDate": {
            "type": "string",
            "example": "2026-10-15",
            "description": "Tanggal panen"
          },
          "scheduleId": {
            "type": "string",
            "example": "uuid-jadwal",
            "description": "ID jadwal tanam terkait (opsional)"
          }
        }
      },
...
    }
  }
}